Move controllers to Refinery::Snippets module; permit params
